Move shrc203_VISADriver.py to hardware folder. Edit import path accordingly.
